+++ crates/awaken-runtime/src/lib.rs
@@ -47,7 +47,7 @@ pub use backend::{
     ExecutionBackendFactory, ExecutionBackendFactoryError, LocalBackend,
 };
 pub use builder::{AgentRuntimeBuilder, BuildError};
-pub use child_agent::{ChildAgentParams, run_child_agent};
+pub use child_agent::{ChildAgentParams, StreamingPassthroughSink, run_child_agent};
 pub use phase::{
     DEFAULT_MAX_PHASE_ROUNDS, ExecutionEnv, PhaseContext, PhaseHook, PhaseRuntime, ToolGateHook,
     ToolPolicyHook, TypedEffectHandler, TypedScheduledActionHandler,
